Move internal models out of public pypi package
