+++ nucliadb/tests/writer/test_reprocess_file_field.py
@@ -24,7 +24,7 @@ import pytest
 from httpx import AsyncClient
 
 from nucliadb.common import datamanagers
-from nucliadb.ingest.processing import ProcessingInfo
+from nucliadb.models.internal.processing import ProcessingInfo
 from nucliadb.writer.api.v1.router import KB_PREFIX, RESOURCE_PREFIX, RESOURCES_PREFIX
 from nucliadb.writer.utilities import get_processing
 from nucliadb_models.resource import QueueType
